move DistrictData class to scrape_common
